+++ API-Produtos.DTO/ProdutosDTO.cs
@@ -2,7 +2,7 @@
 {
     public class ProdutosDTO
     {
-          public int? Id { get; set; } 
+          public long? Id { get; set; } 
           public string nome { get; set; }
           public float valor_unitario { get; set; }
           public int qtde_estoque { get; set; }
